+++ NEWS.md
@@ -1,3 +1,8 @@
+# eatTools 0.3.8.9000
+
+* bugfix in `wideToLong()`
+* new function `do_call_rbind_withName()` allows row binding a list of `data.frames` while keeping their list names 
+
 # eatTools 0.3.8
 
 * `asNumericIfPossible()` now works with matrices